+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* NOTE: APP_DEFAULT_PLANNING_DATA_PROVIDER has been changed to APP_DEFAULT_DATA_PROVIDER. This has to be changed when releasing.
+* [PR-117](https://github.com/itk-dev/economics/pull/117)
+  1211: Added hour report
 * [PR-121](https://github.com/itk-dev/economics/pull/121)
   1485: Fixed floating number issues
 * [PR-120](https://github.com/itk-dev/economics/pull/120)